requirements.txt
ipython
keras
matplotlib
numpy
opencv_contrib_python
opencv_python
Pillow
tensorflow
